Notifications: notify_notification_update would send an empty summary
If summary is not NULL but still an empty string, the following error is
output:

(OSD Lyrics:12344): libnotify-CRITICAL **: notify_notification_update: assertion 'summary != NULL && *summary != '\0'' failed

Enlightenment shows an empty notification when the summary is empty for
the first time, or repeats the previous notification. Now, "Unknown title"
is shown in a proper notification.