daemon: some cleanup in LrcDb
Most of the columns in the database happen to be named following the
METADATA_* consts. METADATA_TRACKNUM is the only exception, and there
seems to be no reason for that since only LrcDb itself uses it.

Simplify things by setting METADATA_TRACKNUM to 'tracknum' like in the
LrcDb column. With this done, we can clarify the code by naming the
columns based on the METADATA_* consts. Note that the database itself is
unchanged by this (this is only a readability improvement), therefore
there is no backwards-compatbility concern.

While at it, create the info query string (which is now much simpler)
directly on class initialisation instead of doing it every time in
_find_by_info().