Delete LrcDb entry upon SetLyricContent()
This commit fixes a number of bugs around the handling of no-lyric
assignments, and also changes the semantics of lyrics file assignments.

This area of the daemon is a bit tricky and took me some time to
understand, so let's start with the current situation. Before this
commit, OSD Lyrics handles lyrics files as such, for a given music file:

* If lyrics are downloaded and the user doesn't take any other action,
  the lyrics will be saved in a file according to the user-defined
  patterns (see LyricsService._save_to_patterns()), and read from there
  when the music file is opened again.

* If an lrc file is assigned ("Assign lyric..." menu option), the lyrics
  will always be read from this file *but also* downloaded lyrics will
  be written to this file. It is possible to assign the music file to
  another lrc file, but the association cannot be undone altogether
  (i.e. removed from the database - LrcDb).

* If the music file is set as "No lyric" (also a menu option), things
  get really bad. If lyrics contents are later downloaded for this file,
  LyricsService.SetLyricContent() will find the no-lyric association in
  the database, and happily try to save the contents to "none:"! This
  makes the daemon very unhappy, since there is no handler to save
  lyrics to "none:". Even if there were, this wouldn't make any
  difference: the contents are not saved anywhere, and therefore won't
  ever appear in the interface.

The 3rd case was clearly broken and needed fixing. Additionally, the 2nd
case made me realise that the semantics of lrc file assignment was not
what I expected. I expected user-assigned lrc files to be viewed as
read-only contents files, not a backend to both read and store lyrics
for the associated music file: if the user then goes on to download
other lyrics, I doubt they would be happy with their original lrc files
(potentially original, written by themselves) being overwritten!

Therefore, in the end, I settled for the most straightforward approach:
when SetLyricContent() is triggered (after lyrics have been downloaded),
any association with the given music file (location + metadata) is
deleted from the database, and the new contents are saved to a generated
path according to the user-defined patterns. This fixes the 3rd case (by
going out of the "none: black hole"), and changes the semantics of the
2nd case by revoking the assignment, thereby preventing the
user-provided file from being overwritten.

There are some other minor fixes in this commit. Notably, a 'none'
handler has been added to _save_to_patterns(), because even though we
don't attempt to save to 'none:' from SetLyricContent() anymore, we
still do from SetOffset(), and we might as well not crash because of
that :-) Ideally the UI shouldn't allow changing the delay at all when
there are no lyrics, but that should do for now.